requirements.txt
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
google-cloud-secret-manager==2.16.4
python-telegram-bot==21.1
openai==1.3.0
schedule==1.2.0
beautifulsoup4==4.12.2
lxml==4.9.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
flask==3.0.0
python-dotenv==1.0.0
